requirements.txt
streamlit>=1.35
openai>=1.40
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.2
pdfplumber>=0.11
scikit-learn>=1.4
pandas>=2.2
langdetect>=1.0.9
joblib>=1.4
